+++ CHANGELOG.md
@@ -18,6 +18,7 @@ You can view the most detailed change log on the https://github.com/embar-/pdsa-
 - Ability to sort the content of preview tables in the 'File Upload' tab (previously only worked in the 'Graph' tab).
 - Ability to filter the content of preview tables in the 'Graph' tab.
 - Ability to copy drawn tables with quotation marks (").
+- Ability to copy only tables selected by mouse (with or without quotes).
 - The ability to export all tables (including their columns and references) to JSON, not just the displayed ones.
 - The user at PDSA can select the column by which tables are filtered for the initial display in the chart 
   (PDSA does not have a separate column for this, but a separate column is available when importing the exported JSON).